A lot of changes in the models, see description
- Item  Prototypes are now loaded from the database only when the
prototype is needed

- Item instance creation is removed from the models, since items are
only fetched as belongings (equipments, inventory, warehouse) or as
drops in the memory. There is no need for any additional item instance
creation.

- The mock model works again.

- The file reader for item prototypes has moved to the mock model

- The inventory and equipments for a player are stored differently

- The equipment slot enumeration was unused due the highly use of native
integer. The enum is removed, static int constansts are added to the
inventory class instead.


- The logic for get ItemPrototype/player/equipmentproto/iteminstance all
moved to AbstractModel, where it should be (not in the specialized
models, these are delegated via the fetch... function calls when an
identity is not existing in the memory).

- The server main class has cleaned up a bit

- Item equip/unequip are now handled differently.

- Bugfix for equiping an item over another item. Item is now correctly
transfered to the inventory.

There still are some issues:
- Model implementations are required (!) to add new instances of item
prototypes, item instances and players to be added to their
corresponding map in the super class (AbstractModel)

- Only equipment instances and equipment prototypes can be instantiated
from the PostgreSQL model. It should somehow differentiate Items and
Equipments.